Remove signing for Tianocore ATF image
The firmware secure update services in the ATF firmware version 1.06x
will abort the capsule update if the firmware image is signed but the
DBU key is not deployed in the system. However, in case the DBU is not
deployed and the image is not signed too, the update is allowed.

This patch removes signing the Tianocore ATF image with the DBU key as
the EDK2 firmware doesn't have an application for DBU deployment.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>